# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel",
        "description": "Receipt API",
        "version": "1.0.1"
    },
    "host": "localhost",
    "basePath": "\/",
    "paths": {
...
        "\/api\/Department": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "schema": {
                            "type": "object",
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ]
            }
        },
        "\/api\/Department\/{Department}": {
            "get": {
                "summary": "Display the specified resource.",
                "description": "",
                "deprecated": false,
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Department",
                        "type": "string",
                        "required": true,
                        "description": ""
                    }
                ]
            },
            "put": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Department",
                        "type": "string",
                        "required": true,
                        "description": ""
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "schema": {
                            "type": "object",
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ]
            },
            "patch": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Department",
                        "type": "string",
                        "required": true,
                        "description": ""
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "schema": {
                            "type": "object",
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ]
            },
            "delete": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Department",
                        "type": "string",
                        "required": true,
                        "description": ""
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "bearerFormat": "JWT",
            "scheme": "bearer"
        }
    }
}